Ensure extras flag is parsed converted to a long instead of an int.

+++ Blish HUD/DebugHelper/Native/MOUSELLHOOKSTRUCT.cs
@@ -9,7 +9,7 @@ namespace Blish_HUD.DebugHelper.Native {
         public int   mouseData;
         public int   flags;
         public int   time;
-        public int   extraInfo;
+        public long  extraInfo;
 
     }
